+++ packages/crdt_lf/lib/src/version_vector/version_vector.dart
@@ -9,7 +9,7 @@ import 'package:hlc_dart/hlc_dart.dart';
 /// Example:
 /// `{client1: HLC(3, 0), client2: HLC(2, 0),client3: HLC(1, 0)}`
 ///
-/// This means that the latest operation for client1 is HLC(3, 0) 
+/// This means that the latest operation for client1 is HLC(3, 0)
 /// (same reasoning for client2 and client3)
 class VersionVector {
   /// Creates a [VersionVector].